refactor: Category - 카테고리 모든조회 API 파일경로 변경 #49
- res , req 디렉토리 생성 및 하위 파일 이동
