+++ src/main/java/com/ioteam/order_management_platform/category/dto/req/CreateCategoryRequestDto.java
@@ -1,4 +1,4 @@
-package com.ioteam.order_management_platform.category.dto;
+package com.ioteam.order_management_platform.category.dto.req;
 
 import org.hibernate.validator.constraints.Length;
